requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
fastmcp>=2.3.0
pydantic>=2.0.0
pydantic-settings>=2.0
google-generativeai>=0.5.0
python-multipart
httpx>=0.24.0
httpx-sse>=0.3.0
sse-starlette>=1.0.0
pymupdf>=1.18.0
itsdangerous>=2.1.2
pymupdf4llm
pyngrok>=7.0.0
requests>=2.28.0
PyYAML>=6.0
aiofiles>=0.8.0